requirements.txt
# =========================
# Cashtube / Digital Asset Ghost Hunter
# Python Requirements
# =========================

# Core HTTP client for YouTube API + future registrar APIs
requests>=2.31.0

# YouTube metadata extraction without full browser automation
yt-dlp>=2024.03.10

# Environment variable loader (.env support)
python-dotenv>=1.0.1

# Optional: Fast CSV / dataframe analysis
pandas>=2.2.2

# Optional: Async HTTP for future Phase 3 scaling
httpx>=0.27.0

# Optional: Retry / resilience
tenacity>=8.2.3

# Optional: PostgreSQL backend (future caching + modular monolith)
psycopg2-binary>=2.9.9

# Optional: OpenTelemetry (future observability)
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0

# Optional: Domain parsing improvements
tldextract>=5.1.2